app/requirements.txt
# Backend - API Framework
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.9

# Database (PostgreSQL + Async)
sqlalchemy[asyncio]>=2.0.25
asyncpg>=0.29.0
psycopg2-binary>=2.9.9
pydantic-settings>=2.1.0
alembic>=1.13.0
geoalchemy2>=0.14.0
shapely>=2.0.0

# Data & Validation
pydantic>=2.6.0
pydantic-core>=2.16.0
numpy>=1.26.0
pandas>=2.2.0

# Security
bcrypt>=4.1.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-dotenv>=1.0.0

# HTTP Client (External APIs)
httpx>=0.27.0
requests>=2.31.0

# ML & AI
scikit-learn>=1.4.0
xgboost>=2.0.0
lightgbm>=4.0.0
scipy>=1.12.0
openai>=1.12.0

# Utilities
email-validator>=2.1.0

# Cache (Redis)
redis>=5.0.0

# Reporting
reportlab>=4.0.0
matplotlib>=3.8.0


# RAG (Retrieval-Augmented Generation)
sentence-transformers>=2.2.0
faiss-cpu>=1.7.0

# Excel & Maps
openpyxl>=3.1.0
xlsxwriter>=3.1.0
openrouteservice>=2.3.0
# FAZ 15 Observability
sentry-sdk>=1.40.0
prometheus-fastapi-instrumentator>=6.1.0
groq>=0.4.0
